MainMenu TopItem Table umgesetzt

+++ components/ILIAS/AdministrativeNotification/classes/class.ilADNNotificationGUI.php
@@ -57,16 +57,32 @@ class ilADNNotificationGUI extends ilADNAbstractGUI
             case self::CMD_CREATE:
                 return $this->create();
             case self::CMD_EDIT:
+                $field = $this->getFieldFromRequest();
+                if ($field === null) {
+                    throw new ilException("Field not found");
+                }
                 return $this->edit();
             case self::CMD_DUPLICATE:
+                $field = $this->getFieldFromRequest();
+                if ($field === null) {
+                    throw new ilException("Field not found");
+                }
                 $this->duplicate();
                 break;
             case self::CMD_UPDATE:
                 return $this->update();
             case self::CMD_DELETE:
+                $field = $this->getFieldFromRequest();
+                if ($field === null) {
+                    throw new ilException("Field not found");
+                }
                 $this->delete();
                 break;
             case self::CMD_RESET:
+                $field = $this->getFieldFromRequest();
+                if ($field === null) {
+                    throw new ilException("Field not found");
+                }
                 $this->reset();
                 break;
             case self::CMD_DEFAULT:
@@ -87,9 +103,22 @@ class ilADNNotificationGUI extends ilADNAbstractGUI
             );
             $this->toolbar->addComponent($btn_add_msg);
         }
-
-        //return (new ilADNNotificationTableGUI($this, self::CMD_DEFAULT))->getHTML();
         return (new \ILIAS\AdministrativeNotification\Table($this))->getHTML();
+    }
+
+    protected function getFieldIdFromRequest(): int
+    {
+        $query_params = $this->http->request()->getQueryParams(); // aka $_GET
+        $name = $this->table->getIdToken()->getName(); // name of the query parameter from the table
+        $field_ids = $query_params[$name] ?? []; // array of field ids
+        return (int) (is_array($field_ids) ? end($field_ids) : $field_ids); // return the last field id
+    }
+
+    private function getFieldFromRequest(): ActiveRecord
+    {
+        $field_id = $this->getFieldIdFromRequest();
+
+        return ilADNNotification::findOrFail($field_id); // get field from id
     }
 
     protected function add(): string
